+++ crates/teloxide-core/src/bot/download.rs
@@ -1,6 +1,12 @@
+use std::path::Path;
+
 use bytes::Bytes;
 use futures::{future::BoxFuture, stream::BoxStream, FutureExt, StreamExt};
-use tokio::io::AsyncWrite;
+
+use tokio::{
+    fs::File,
+    io::{AsyncReadExt, AsyncWrite, AsyncWriteExt},
+};
 
 use crate::{
     bot::Bot,
@@ -17,9 +23,19 @@ impl Download for Bot {
 
     fn download_file<'dst>(
         &self,
-        path: &str,
+        path: &'dst str,
         destination: &'dst mut (dyn AsyncWrite + Unpin + Send),
     ) -> Self::Fut<'dst> {
+        let is_localhost = match &self.api_url.host_str() {
+            Some(host) => ["localhost", "127.0.0.1"].contains(host),
+            None => false,
+        };
+        // If path is absolute and api_url contains localhost, it is pretty clear there
+        // is a local TBA server with --local option, we can just copy the file
+        if is_localhost && Path::new(&path).is_absolute() {
+            return copy_file(path, destination).boxed();
+        }
+
         net::download_file(
             &self.client,
             reqwest::Url::clone(&*self.api_url),
@@ -44,4 +60,24 @@ impl Download for Bot {
         .map(|res| res.map_err(crate::errors::hide_token))
         .boxed()
     }
+}
+
+async fn copy_file<'o, D>(path: &'o str, dst: &'o mut D) -> Result<(), DownloadError>
+where
+    D: ?Sized + AsyncWrite + Unpin,
+{
+    let mut src_file = File::open(path).await?;
+
+    let mut buffer = [0; 1024];
+    loop {
+        let n = src_file.read(&mut buffer).await?;
+        if n == 0 {
+            break;
+        }
+
+        dst.write_all(&buffer[..n]).await?;
+    }
+
+    dst.flush().await?;
+    Ok(())
 }